requirements.txt
scipy==1.5.0
numpy==1.19.0
soundfile==0.10.3.post1
python-levenshtein==0.12.0
torch==1.5.1
torchelastic==0.2.0
visdom==0.1.8.9
wget==3.2
librosa==0.7.2
numba==0.43.0
llvmlite==0.32.1
tqdm==4.47.0
matplotlib==3.2.2
flask==1.1.2
sox==1.3.7
pandas==1.0.5
scikit-learn==0.23.1
hydra-core==1.0.0rc1
comet-ml==3.1.12
pytorch-lightning==0.8.5
